requirements.txt
# ─────────────────────────────────────────────────────────────
# TalentLens — pinned, mutually-compatible dependency set
# Python 3.10  |  tested 2026-04-23
#
# Install:  pip install -r requirements.txt
# After install run:
#   pip install "protobuf>=4.0.0,<5.0.0"   ← locks the protobuf sweet-spot
#   (mlflow needs <5, dbt 1.9 needs >=4 — this one command satisfies both)
# ─────────────────────────────────────────────────────────────

# ── Data / ETL ───────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
faker==24.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
alembic==1.13.1
python-dotenv==1.0.1
pyarrow==15.0.2

# ── Kafka ────────────────────────────────────────────────────
kafka-python==2.0.2

# ── API ──────────────────────────────────────────────────────
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
httpx==0.27.0

# ── dbt  (upgraded from 1.7.4 → 1.9.0 for protobuf compat) ──
# Install separately:  pip install dbt-postgres==1.9.0
# (dbt pulls in its own dep tree; listing it here causes resolver conflicts)

# ── ML (Phase 3) ─────────────────────────────────────────────
xgboost==2.0.3
scikit-learn==1.4.2
shap==0.45.0
mlflow==2.12.1
optuna==3.6.1
fairlearn==0.10.0

# ── Protobuf — MUST be pinned after all other installs ───────
# mlflow 2.12.1 requires <6 ; dbt-core 1.9 requires >=4
# Run:  pip install "protobuf>=4.0.0,<5.0.0"
# (not pinned here to avoid fighting the resolver on first install)

# ── LLM (Phase 5) ────────────────────────────────────────────
langchain==0.1.16
langchain-openai==0.1.3
langchain-anthropic==0.1.11
langchain-google-genai==1.0.3
langgraph>=0.0.30
faiss-cpu==1.8.0
sentence-transformers==2.7.0
openai==1.23.1
anthropic==0.25.0
google-generativeai==0.5.4

# ── PySpark (Phase 2) ────────────────────────────────────────
pyspark==3.5.1
# Requires Java 11 or 17.  On this Mac, Java lives inside Anaconda:
#   export JAVA_HOME=/opt/anaconda3
# To persist across sessions (run once):
#   mkdir -p /opt/anaconda3/envs/talentlens/etc/conda/activate.d
#   echo 'export JAVA_HOME=/opt/anaconda3' \
#     > /opt/anaconda3/envs/talentlens/etc/conda/activate.d/java_home.sh

# ── Observability ────────────────────────────────────────────
prometheus-client==0.20.0

# ── Testing ──────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
locust==2.24.1
